refactor: move calculationBucket to own file
