--- src/general_manager/interface/__init__.py
@@ -1,4 +1,0 @@
-# from .databaseInterface import DatabaseInterface, ReadOnlyInterface, DatabaseBucket
-# from .calculationInterface import CalculationInterface, CalculationBucket
-# from .readOnlyInterface import ReadOnlyInterface
-# from .databaseInterface import DatabaseInterface
